requirements.txt
asgiref==3.5.2
Django==4.1.3
flake8==5.0.4
mccabe==0.7.0
psycopg2==2.9.5
pycodestyle==2.9.1
pyflakes==2.5.0
python-dotenv==0.21.0
sqlparse==0.4.3
